+++ packages/expo-modules-core/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Ensure `uuid.v4`and`uuid.v5` is available on old react native architecture. ([#33621](https://github.com/expo/expo/pull/33621) by [@andrejpavlovic](https://github.com/andrejpavlovic))
 - Changed `import` to `import type` for TS type declarations. ([#33447](https://github.com/expo/expo/pull/33447) by [@j-piasecki](https://github.com/j-piasecki))
 - [macOS] Allow SwiftUI views to work on macOS ([#33506](https://github.com/expo/expo/pull/33506) by [@hassankhan](https://github.com/hassankhan))
+- [iOS] Fix `ExpoAppDelegate` lifecycle events with scenes ([#33691](https://github.com/expo/expo/pull/33691) by [@alextoudic](https://github.com/alextoudic))
 
 ### 💡 Others